tgbot_service/requirements.txt
uvicorn
httpx
pydantic
datetime
fastapi
pydantic_settings
python-telegram-bot==22.4
requests
python-dotenv